Fix test, add CI, forward ports, cleanups (#3)
Before this change, our test was broken. With this change, we fix the
test, and introduce CI to make sure it stays fixed.

This change also sets the dev container to forward both ports needed to
reach the application and its frontend.

It also fixes a few typos, adds a few comments, and cleans up a few
pieces of unused code.